+++ contracts/src/mocks/plugin/PluginCloneableSetupMock.sol
@@ -6,6 +6,7 @@ pragma solidity ^0.8.8;
 import {PermissionLib} from "../../permission/PermissionLib.sol";
 import {IPluginSetup} from "../../plugin/setup/IPluginSetup.sol";
 import {PluginSetup} from "../../plugin/setup/PluginSetup.sol";
+import {ProxyLib} from "../../utils/deployment/ProxyLib.sol";
 import {IDAO} from "../../dao/IDAO.sol";
 import {mockPermissions, mockHelpers} from "./PluginSetupMockData.sol";
 import {PluginCloneableMockBuild1, PluginCloneableMockBuild2} from "./PluginCloneableMock.sol";
@@ -14,35 +15,34 @@ import {PluginCloneableMockBuild1, PluginCloneableMockBuild2} from "./PluginClon
 /// v1.1 (Release 1, Build 1)
 /// @dev DO NOT USE IN PRODUCTION!
 contract PluginCloneableSetupMockBuild1 is PluginSetup {
-    address internal pluginBase;
+    using ProxyLib for address;
 
-    constructor() {
-        pluginBase = address(new PluginCloneableMockBuild1());
-    }
+    uint16 internal constant THIS_BUILD = 1;
+
+    constructor() PluginSetup(address(new PluginCloneableMockBuild1())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) external override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) external returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(PluginCloneableMockBuild1.initialize, (IDAO(_dao)));
-        plugin = createERC1967Proxy(pluginBase, initData); // TODO createClone(pluginBase, initData); is missing! See task OS-794 and OS-675.
-        preparedSetupData.helpers = mockHelpers(1);
-        preparedSetupData.permissions = mockPermissions(0, 1, PermissionLib.Operation.Grant);
+        plugin = implementation().deployMinimalProxy(initData);
+        preparedSetupData.helpers = mockHelpers(THIS_BUILD);
+        preparedSetupData.permissions = mockPermissions(
+            0,
+            THIS_BUILD,
+            PermissionLib.Operation.Grant
+        );
     }
 
     /// @inheritdoc IPluginSetup
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external pure override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
-        permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
-        return address(pluginBase);
+        permissions = mockPermissions(0, THIS_BUILD, PermissionLib.Operation.Revoke);
     }
 }
 
@@ -50,34 +50,33 @@ contract PluginCloneableSetupMockBuild1 is PluginSetup {
 /// v1.2 (Release 1, Build 2)
 /// @dev DO NOT USE IN PRODUCTION!
 contract PluginCloneableSetupMockBuild2 is PluginSetup {
-    address internal pluginBase;
+    using ProxyLib for address;
 
-    constructor() {
-        pluginBase = address(new PluginCloneableMockBuild2());
-    }
+    uint16 internal constant THIS_BUILD = 2;
+
+    constructor() PluginSetup(address(new PluginCloneableMockBuild2())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) external override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) external returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(PluginCloneableMockBuild2.initialize, (IDAO(_dao)));
-        plugin = createERC1967Proxy(pluginBase, initData); // TODO createClone(pluginBase, initData); is missing! See task OS-794 and OS-675.
-        preparedSetupData.helpers = mockHelpers(2);
-        preparedSetupData.permissions = mockPermissions(0, 2, PermissionLib.Operation.Grant);
+        plugin = implementation().deployMinimalProxy(initData);
+        preparedSetupData.helpers = mockHelpers(THIS_BUILD);
+        preparedSetupData.permissions = mockPermissions(
+            0,
+            THIS_BUILD,
+            PermissionLib.Operation.Grant
+        );
     }
 
     /// @inheritdoc IPluginSetup
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external pure override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
-        permissions = mockPermissions(0, 2, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
-        return address(pluginBase);
+        permissions = mockPermissions(0, THIS_BUILD, PermissionLib.Operation.Revoke);
     }
 }