+++ contracts/src/mocks/plugin/PluginSetupMock.sol
@@ -14,17 +14,13 @@ import {PluginMockBuild1} from "./PluginMock.sol";
 /// v1.1 (Release 1, Build 1)
 /// @dev DO NOT USE IN PRODUCTION!
 contract PluginSetupMockBuild1 is PluginSetup {
-    address internal pluginBase;
-
-    constructor() {
-        pluginBase = address(new PluginMockBuild1(IDAO(address(0))));
-    }
+    constructor() PluginSetup(address(new PluginMockBuild1(IDAO(address(0))))) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) external override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) external returns (address plugin, PreparedSetupData memory preparedSetupData) {
         plugin = address(new PluginMockBuild1(IDAO(_dao)));
         preparedSetupData.helpers = mockHelpers(1);
         preparedSetupData.permissions = mockPermissions(0, 1, PermissionLib.Operation.Grant);
@@ -34,13 +30,8 @@ contract PluginSetupMockBuild1 is PluginSetup {
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external pure override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
-        return address(pluginBase);
     }
 }