+++ contracts/src/mocks/plugin/PluginUUPSUpgradeableSetupMock.sol
@@ -5,7 +5,8 @@ pragma solidity ^0.8.8;
 
 import {PermissionLib} from "../../permission/PermissionLib.sol";
 import {IPluginSetup} from "../../plugin/setup/IPluginSetup.sol";
-import {PluginSetup} from "../../plugin/setup/PluginSetup.sol";
+import {PluginUpgradeableSetup} from "../../plugin/setup/PluginUpgradeableSetup.sol";
+import {ProxyLib} from "../../utils/deployment/ProxyLib.sol";
 import {IDAO} from "../../dao/IDAO.sol";
 import {mockPermissions, mockHelpers} from "./PluginSetupMockData.sol";
 import {PluginUUPSUpgradeableMockBuild1, PluginUUPSUpgradeableMockBuild2, PluginUUPSUpgradeableMockBuild3} from "./PluginUUPSUpgradeableMock.sol";
@@ -13,70 +14,80 @@ import {PluginUUPSUpgradeableMockBuild1, PluginUUPSUpgradeableMockBuild2, Plugin
 /// @notice A mock plugin setup of an upgradeable plugin to be deployed via the UUPS pattern.
 /// v1.1 (Release 1, Build 1)
 /// @dev DO NOT USE IN PRODUCTION!
-contract PluginUUPSUpgradeableSetupMockBuild1 is PluginSetup {
-    address internal pluginBase;
+contract PluginUUPSUpgradeableSetupMockBuild1 is PluginUpgradeableSetup {
+    using ProxyLib for address;
 
-    constructor() {
-        pluginBase = address(new PluginUUPSUpgradeableMockBuild1());
-    }
+    uint16 internal constant THIS_BUILD = 1;
+
+    constructor() PluginUpgradeableSetup(address(new PluginUUPSUpgradeableMockBuild1())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) public returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(
             PluginUUPSUpgradeableMockBuild1.initialize,
             (IDAO(_dao))
         );
-        plugin = createERC1967Proxy(pluginBase, initData);
+        plugin = implementation().deployUUPSProxy(initData);
         preparedSetupData.helpers = mockHelpers(1);
         preparedSetupData.permissions = mockPermissions(0, 1, PermissionLib.Operation.Grant);
+    }
+
+    /// @inheritdoc IPluginSetup
+    /// @dev The default implementation for the initial build 1 that reverts because no earlier build exists.
+    function prepareUpdate(
+        address _dao,
+        uint16 _fromBuild,
+        SetupPayload calldata _payload
+    ) external pure virtual returns (bytes memory, PreparedSetupData memory) {
+        (_dao, _fromBuild, _payload);
+        revert InvalidUpdatePath({fromBuild: 0, thisBuild: THIS_BUILD});
     }
 
     /// @inheritdoc IPluginSetup
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
-        permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view virtual override returns (address) {
-        return address(pluginBase);
+        permissions = mockPermissions(0, THIS_BUILD, PermissionLib.Operation.Revoke);
     }
 }
 
 /// @notice A mock plugin setup of an upgradeable plugin to be deployed via the UUPS pattern.
 /// v1.2 (Release 1, Build 2)
 /// @dev DO NOT USE IN PRODUCTION!
-contract PluginUUPSUpgradeableSetupMockBuild2 is PluginSetup {
-    address internal pluginBase;
+contract PluginUUPSUpgradeableSetupMockBuild2 is PluginUpgradeableSetup {
+    using ProxyLib for address;
 
-    constructor() {
-        pluginBase = address(new PluginUUPSUpgradeableMockBuild2());
-    }
+    uint16 internal constant THIS_BUILD = 2;
+
+    constructor() PluginUpgradeableSetup(address(new PluginUUPSUpgradeableMockBuild2())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
-    ) external override returns (address plugin, PreparedSetupData memory preparedSetupData) {
+    ) external returns (address plugin, PreparedSetupData memory preparedSetupData) {
         bytes memory initData = abi.encodeCall(
             PluginUUPSUpgradeableMockBuild2.initialize,
             (IDAO(_dao))
         );
-        plugin = createERC1967Proxy(pluginBase, initData);
-        preparedSetupData.helpers = mockHelpers(2);
-        preparedSetupData.permissions = mockPermissions(0, 2, PermissionLib.Operation.Grant);
+        plugin = implementation().deployUUPSProxy(initData);
+        preparedSetupData.helpers = mockHelpers(THIS_BUILD);
+        preparedSetupData.permissions = mockPermissions(
+            0,
+            THIS_BUILD,
+            PermissionLib.Operation.Grant
+        );
     }
 
     /// @inheritdoc IPluginSetup
     function prepareUpdate(
         address _dao,
-        uint16 _currentBuild,
+        uint16 _fromBuild,
         SetupPayload calldata _payload
     )
         external
@@ -87,13 +98,14 @@ contract PluginUUPSUpgradeableSetupMockBuild2 is PluginSetup {
         (_dao, _payload);
 
         // Update from Build 1
-        if (_currentBuild == 1) {
-            preparedSetupData.helpers = mockHelpers(2);
-            initData = abi.encodeCall(
-                PluginUUPSUpgradeableMockBuild2.initializeFrom,
-                (_currentBuild)
+        if (_fromBuild == 1) {
+            preparedSetupData.helpers = mockHelpers(THIS_BUILD);
+            initData = abi.encodeCall(PluginUUPSUpgradeableMockBuild2.initializeFrom, (_fromBuild));
+            preparedSetupData.permissions = mockPermissions(
+                _fromBuild,
+                THIS_BUILD,
+                PermissionLib.Operation.Grant
             );
-            preparedSetupData.permissions = mockPermissions(1, 2, PermissionLib.Operation.Grant);
         }
     }
 
@@ -101,25 +113,20 @@ contract PluginUUPSUpgradeableSetupMockBuild2 is PluginSetup {
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external pure override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
-        permissions = mockPermissions(0, 2, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
-        return address(pluginBase);
+        permissions = mockPermissions(0, THIS_BUILD, PermissionLib.Operation.Revoke);
     }
 }
 
 /// @notice A mock plugin setup of an upgradeable plugin to be deployed via the UUPS pattern.
 /// v1.3 (Release 1, Build 3)
-contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
-    address internal pluginBase;
+contract PluginUUPSUpgradeableSetupMockBuild3 is PluginUpgradeableSetup {
+    using ProxyLib for address;
 
-    constructor() {
-        pluginBase = address(new PluginUUPSUpgradeableMockBuild3());
-    }
+    uint16 internal constant THIS_BUILD = 3;
+
+    constructor() PluginUpgradeableSetup(address(new PluginUUPSUpgradeableMockBuild3())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
@@ -130,15 +137,20 @@ contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
             PluginUUPSUpgradeableMockBuild3.initialize,
             (IDAO(_dao))
         );
-        plugin = createERC1967Proxy(pluginBase, initData);
-        preparedSetupData.helpers = mockHelpers(3);
-        preparedSetupData.permissions = mockPermissions(0, 3, PermissionLib.Operation.Grant);
+        plugin = implementation().deployUUPSProxy(initData);
+        preparedSetupData.helpers = mockHelpers(THIS_BUILD);
+        preparedSetupData.permissions = mockPermissions(
+            0,
+            THIS_BUILD,
+            PermissionLib.Operation.Grant
+        );
     }
 
     /// @inheritdoc IPluginSetup
+
     function prepareUpdate(
         address _dao,
-        uint16 _currentBuild,
+        uint16 _fromBuild,
         SetupPayload calldata _payload
     )
         external
@@ -149,23 +161,28 @@ contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
         (_dao, _payload);
 
         // Update from Build 1
-        if (_currentBuild == 1) {
-            preparedSetupData.helpers = mockHelpers(3);
-            initData = abi.encodeCall(
-                PluginUUPSUpgradeableMockBuild3.initializeFrom,
-                (_currentBuild)
+        if (_fromBuild == 1) {
+            preparedSetupData.helpers = mockHelpers(THIS_BUILD);
+            initData = abi.encodeCall(PluginUUPSUpgradeableMockBuild3.initializeFrom, (_fromBuild));
+            preparedSetupData.permissions = mockPermissions(
+                _fromBuild,
+                THIS_BUILD,
+                PermissionLib.Operation.Grant
             );
-            preparedSetupData.permissions = mockPermissions(1, 3, PermissionLib.Operation.Grant);
+
+            // If this update path should not be supported, you can revert with an error, e.g.,
+            // revert InvalidUpdatePath({fromBuild: _fromBuild, thisBuild: THIS_BUILD});
         }
 
         // Update from Build 2
-        if (_currentBuild == 2) {
-            preparedSetupData.helpers = mockHelpers(3);
-            initData = abi.encodeCall(
-                PluginUUPSUpgradeableMockBuild3.initializeFrom,
-                (_currentBuild)
+        if (_fromBuild == 2) {
+            preparedSetupData.helpers = mockHelpers(THIS_BUILD);
+            initData = abi.encodeCall(PluginUUPSUpgradeableMockBuild3.initializeFrom, (_fromBuild));
+            preparedSetupData.permissions = mockPermissions(
+                _fromBuild,
+                THIS_BUILD,
+                PermissionLib.Operation.Grant
             );
-            preparedSetupData.permissions = mockPermissions(2, 3, PermissionLib.Operation.Grant);
         }
     }
 
@@ -173,13 +190,8 @@ contract PluginUUPSUpgradeableSetupMockBuild3 is PluginSetup {
     function prepareUninstallation(
         address _dao,
         SetupPayload calldata _payload
-    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+    ) external pure returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
-        permissions = mockPermissions(0, 3, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view override returns (address) {
-        return address(pluginBase);
+        permissions = mockPermissions(0, THIS_BUILD, PermissionLib.Operation.Revoke);
     }
 }