+++ contracts/src/plugin/setup/IPluginSetup.sol
@@ -42,13 +42,13 @@ interface IPluginSetup {
 
     /// @notice Prepares the update of a plugin.
     /// @param _dao The address of the updating DAO.
-    /// @param _currentBuild The build number of the plugin to update from.
+    /// @param _fromBuild The build number of the plugin to update from.
     /// @param _payload The relevant data necessary for the `prepareUpdate`. See above.
     /// @return initData The initialization data to be passed to upgradeable contracts when the update is applied in the `PluginSetupProcessor`.
     /// @return preparedSetupData The deployed plugin's relevant data which consists of helpers and permissions.
     function prepareUpdate(
         address _dao,
-        uint16 _currentBuild,
+        uint16 _fromBuild,
         SetupPayload calldata _payload
     ) external returns (bytes memory initData, PreparedSetupData memory preparedSetupData);
 
@@ -63,6 +63,6 @@ interface IPluginSetup {
 
     /// @notice Returns the plugin implementation address.
     /// @return The address of the plugin implementation contract.
-    /// @dev The implementation can be instantiated via the `new` keyword, cloned via the minimal clones pattern (see [ERC-1167](https://eips.ethereum.org/EIPS/eip-1167)), or proxied via the UUPS pattern (see [ERC-1822](https://eips.ethereum.org/EIPS/eip-1822)).
+    /// @dev The implementation can be instantiated via the `new` keyword, cloned via the minimal proxy pattern (see [ERC-1167](https://eips.ethereum.org/EIPS/eip-1167)), or proxied via the UUPS proxy pattern (see [ERC-1822](https://eips.ethereum.org/EIPS/eip-1822)).
     function implementation() external view returns (address);
 }