+++ contracts/src/plugin/setup/PluginUpgradeableSetup.sol
@@ -8,34 +8,24 @@ import {IProtocolVersion} from "../../utils/versioning/IProtocolVersion.sol";
 import {ProtocolVersion} from "../../utils/versioning/ProtocolVersion.sol";
 import {IPluginSetup} from "./IPluginSetup.sol";
 
-/// @title PluginSetup
+/// @title PluginUpgradeableSetup
 /// @author Aragon Association - 2022-2024
-/// @notice An abstract contract to inherit from to implement the plugin setup for non-upgradeable plugins, i.e,
-/// - `Plugin` being deployed via the `new` keyword
-/// - `PluginCloneable` being deployed via the minimal proxy pattern (see [ERC-1167](https://eips.ethereum.org/EIPS/eip-1167)).
+/// @notice An abstract contract to inherit from to implement the plugin setup for upgradeable plugins, i.e, `PluginUUPSUpgradeable` being deployed via the UUPS pattern (see [ERC-1822](https://eips.ethereum.org/EIPS/eip-1822) and [ERC-1967](https://eips.ethereum.org/EIPS/eip-1967)).
 /// @custom:security-contact sirt@aragon.org
-abstract contract PluginSetup is ERC165, IPluginSetup, ProtocolVersion {
-    /// @notice The address of the plugin implementation contract for initial block explorer verification and, in the case of `PluginClonable` implementations, to create [ERC-1167](https://eips.ethereum.org/EIPS/eip-1167) clones from.
+abstract contract PluginUpgradeableSetup is ERC165, IPluginSetup, ProtocolVersion {
+    /// @notice The address of the plugin implementation contract for initial block explorer verification
+    /// and to create [ERC-1967](https://eips.ethereum.org/EIPS/eip-1967) UUPS proxies from.
     address internal immutable IMPLEMENTATION;
 
-    /// @notice Thrown when attempting to prepare an update on a non-upgradeable plugin.
-    error NonUpgradeablePlugin();
+    /// @notice Thrown when an update path is not available, for example, if this is the initial build.
+    /// @param fromBuild The build number to update from.
+    /// @param thisBuild The build number of this setup to update to.
+    error InvalidUpdatePath(uint16 fromBuild, uint16 thisBuild);
 
     /// @notice The contract constructor, that setting the plugin implementation contract.
     /// @param _implementation The address of the plugin implementation contract.
     constructor(address _implementation) {
         IMPLEMENTATION = _implementation;
-    }
-
-    /// @inheritdoc IPluginSetup
-    /// @dev Since the underlying plugin is non-upgradeable, this non-virtual function must always revert.
-    function prepareUpdate(
-        address _dao,
-        uint16 _fromBuild,
-        SetupPayload calldata _payload
-    ) external pure returns (bytes memory, PreparedSetupData memory) {
-        (_dao, _fromBuild, _payload);
-        revert NonUpgradeablePlugin();
     }
 
     /// @notice Checks if this or the parent contract supports an interface by its ID.