+++ contracts/test/plugin/plugin-clonable.ts
@@ -1,37 +1,42 @@
 import {
-  CloneFactory__factory,
+  DAOMock,
+  DAOMock__factory,
   IPlugin__factory,
   IProtocolVersion__factory,
+  PluginCloneableMockBad__factory,
   PluginCloneableMockBuild1,
   PluginCloneableMockBuild1__factory,
+  ProxyFactory__factory,
 } from '../../typechain';
+import {
+  ProxyCreatedEvent,
+  ProxyFactory,
+} from '../../typechain/src/utils/deployment/ProxyFactory';
 import {erc165ComplianceTests, getOzInitializedSlotValue} from '../helpers';
 import {osxCommonsContractsVersion} from '../utils/versioning/protocol-version';
-import {getInterfaceId, PluginType} from '@aragon/osx-commons-sdk';
+import {
+  findEvent,
+  getInterfaceId,
+  PluginType,
+  PROXY_FACTORY_EVENTS,
+} from '@aragon/osx-commons-sdk';
+import {loadFixture} from '@nomicfoundation/hardhat-network-helpers';
 import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 import {expect} from 'chai';
 import {ethers} from 'hardhat';
 
 describe('PluginCloneable', function () {
-  let deployer: SignerWithAddress;
-  let plugin: PluginCloneableMockBuild1;
-
-  before(async () => {
-    [deployer] = await ethers.getSigners();
-    plugin = await new PluginCloneableMockBuild1__factory(deployer).deploy();
-  });
-
   describe('Initializable', async () => {
     it('initialize', async () => {
-      // Deploy a clone
-      const cloneFactory = await new CloneFactory__factory(deployer).deploy(
-        plugin.address
+      const {proxyFactory, Build1Factory, daoMock} = await loadFixture(fixture);
+
+      // Deploy an uninitialized clone
+      const tx = await proxyFactory.deployMinimalProxy([]);
+      const event = await findEvent<ProxyCreatedEvent>(
+        tx,
+        PROXY_FACTORY_EVENTS.ProxyCreated
       );
-      const expectedAddress = await cloneFactory.callStatic.deployClone();
-      await cloneFactory.deployClone();
-      const clone = new PluginCloneableMockBuild1__factory(deployer).attach(
-        expectedAddress
-      );
+      const clone = Build1Factory.attach(event.args.proxy);
 
       // Check the clone before initialization
       expect(
@@ -40,28 +45,54 @@ describe('PluginCloneable', function () {
       expect(await clone.dao()).to.equal(ethers.constants.AddressZero);
       expect(await clone.state1()).to.equal(0);
 
-      // Initialize the clone
-      const dummyDaoAddress = plugin.address;
-      await clone.initialize(dummyDaoAddress);
+      await clone.initialize(daoMock.address);
 
       // Check the clone after initialization
       expect(
         await getOzInitializedSlotValue(ethers.provider, clone.address)
       ).to.equal(1);
-      expect(await clone.dao()).to.equal(dummyDaoAddress);
+      expect(await clone.dao()).to.equal(daoMock.address);
       expect(await clone.state1()).to.equal(1);
+    });
+
+    it('disables initializers for the implementation', async () => {
+      const {implementation} = await loadFixture(fixture);
+
+      // Check that the implementation is uninitialized.
+      expect(await implementation.dao()).to.equal(ethers.constants.AddressZero);
+      expect(await implementation.state1()).to.equal(0);
+
+      // Check that the implementation initialization is disabled.
+
+      expect(
+        await getOzInitializedSlotValue(ethers.provider, implementation.address)
+      ).to.equal(255);
+    });
+
+    it('reverts if an function tries to call `__PluginCloneable_init` without being an initializer', async () => {
+      const {deployer, daoMock} = await loadFixture(fixture);
+
+      const badPlugin = await new PluginCloneableMockBad__factory(
+        deployer
+      ).deploy();
+
+      await expect(
+        badPlugin.notAnInitializer(daoMock.address)
+      ).to.be.revertedWith('Initializable: contract is not initializing');
     });
   });
 
   describe('PluginType', async () => {
     it('returns the current protocol version', async () => {
-      expect(await plugin.pluginType()).to.equal(PluginType.Cloneable);
+      const {implementation} = await loadFixture(fixture);
+      expect(await implementation.pluginType()).to.equal(PluginType.Cloneable);
     });
   });
 
   describe('ProtocolVersion', async () => {
     it('returns the current protocol version matching the semantic version of the `osx-contracts-commons` package', async () => {
-      expect(await plugin.protocolVersion()).to.deep.equal(
+      const {implementation} = await loadFixture(fixture);
+      expect(await implementation.protocolVersion()).to.deep.equal(
         osxCommonsContractsVersion()
       );
     });
@@ -69,25 +100,67 @@ describe('PluginCloneable', function () {
 
   describe('ERC-165', async () => {
     it('supports the `ERC-165` standard', async () => {
-      await erc165ComplianceTests(plugin, deployer);
+      const {deployer, implementation} = await loadFixture(fixture);
+      await erc165ComplianceTests(implementation, deployer);
     });
 
     it('supports the `IPlugin` interface', async () => {
+      const {implementation} = await loadFixture(fixture);
       const iface = IPlugin__factory.createInterface();
-      expect(await plugin.supportsInterface(getInterfaceId(iface))).to.be.true;
+      expect(await implementation.supportsInterface(getInterfaceId(iface))).to
+        .be.true;
     });
 
     it('supports the `IProtocolVersion` interface', async () => {
+      const {implementation} = await loadFixture(fixture);
       const iface = IProtocolVersion__factory.createInterface();
-      expect(await plugin.supportsInterface(getInterfaceId(iface))).to.be.true;
+      expect(await implementation.supportsInterface(getInterfaceId(iface))).to
+        .be.true;
     });
   });
 
   describe('Protocol version', async () => {
     it('returns the current protocol version', async () => {
-      expect(await plugin.protocolVersion()).to.deep.equal(
+      const {implementation} = await loadFixture(fixture);
+      expect(await implementation.protocolVersion()).to.deep.equal(
         osxCommonsContractsVersion()
       );
     });
   });
 });
+
+type FixtureResult = {
+  deployer: SignerWithAddress;
+  implementation: PluginCloneableMockBuild1;
+  proxyFactory: ProxyFactory;
+  daoMock: DAOMock;
+  initCalldata: string;
+  Build1Factory: PluginCloneableMockBuild1__factory;
+};
+
+async function fixture(): Promise<FixtureResult> {
+  const [deployer] = await ethers.getSigners();
+
+  const Build1Factory = new PluginCloneableMockBuild1__factory(deployer);
+  const daoMock = await new DAOMock__factory(deployer).deploy();
+
+  const implementation = await Build1Factory.deploy();
+
+  const proxyFactory = await new ProxyFactory__factory(deployer).deploy(
+    implementation.address
+  );
+
+  const initCalldata = implementation.interface.encodeFunctionData(
+    'initialize',
+    [daoMock.address]
+  );
+
+  return {
+    deployer,
+    implementation,
+    proxyFactory,
+    daoMock,
+    initCalldata,
+    Build1Factory,
+  };
+}