+++ contracts/test/utils/deployment/clone-factory.ts
@@ -1,1 +1,0 @@
-// TODO CloneFactory will be refactored as part of Task OS-794.